assist/python/requirements-1.txt
about-time==4.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
akshare==1.14.44
alive-progress==3.1.5
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
autopep8==2.3.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.5
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.10
DataRecorder==3.5.3
ddt==1.7.2
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
DownloadKit==2.0.2
DrissionPage==4.0.5.6
dukpy==0.4.0
dynaconf==3.2.6
et-xmlfile==1.1.0
executing==2.0.1
Faker==26.0.0
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.4.1
grapheme==0.6.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huaweicloudsdkcore==3.1.107
huaweicloudsdkocr==3.1.107
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
imgaug==0.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
iso3166==2.1.1
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpath==0.82.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
libretranslatepy==2.1.1
lmdb==1.5.1
logging==0.4.9.6
loguru==0.7.2
lux==0.5.1
lux-api==0.5.1
lux-widget==0.1.11
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mini-racer==0.12.4
mistune==3.0.2
MouseInfo==0.1.3
moviepy==1.0.3
mpy==0.1
multidict==6.0.5
mutagen==1.47.0
mutf8==1.0.6
mypy==1.11.1
mypy-extensions==1.0.0
narwhals==1.6.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
office==0.0.1a0
openai==1.37.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pathlib==1.0.1
pdf2docx==0.5.8
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi==0.1.2
pikepdf==9.1.0
pillow==10.4.0
platformdirs==4.2.2
poai==0.0.11
pocode==0.0.3
poemail==0.0.4
poexcel==0.0.22
pofile==0.1.3
poimage==0.0.22
poocr==0.1.15
popdf==0.1.2
poppt==0.1.1
poprogress==0.0.2
porobot==0.0.3
povideo==0.0.7
poword==0.0.16
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyEmail==0.0.1
PyExecJS==1.5.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyOfficeRobot==0.1.20
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-office==0.4.23
python-pptx==0.6.23
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywifi==1.1.12
pywin32==306
pywinauto==0.6.8
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.5
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
search4file==0.1.15
selenium==4.23.1
Send2Trash==1.8.3
setuptools==72.2.0
sh==2.0.7
shapely==2.0.5
shiboken6==6.7.2
SimpleGUICS2Pygame==2.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
snownlp==0.12.3
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
stack-data==0.6.3
tabulate==0.9.0
tencentcloud-sdk-python==3.0.1200
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
tldextract==5.1.2
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
translate==3.6.1
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
uiautomation==2.0.20
uri-template==1.3.0
urllib3==2.2.2
w3lib==2.2.1
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
wftools==0.0.9
widgetsnbextension==4.0.13
win32-setctime==1.1.0
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
wxPython==4.2.1
xlrd==1.2.0
XlsxWriter==3.2.0
xlwings==0.31.10
xlwt==1.3.0
yarl==1.9.4
you-get==0.4.1718
yt-dlp==2024.8.6
zhipuai==2.1.2
